requirements.txt
pandas
numpy
google-generativeai
streamlit
scikit-learn
PyPDF2
python-dotenv>=0.19.0
#-e .